Implement model validation with Validator interface
- I have removed the IsCompleted member from the Validator interface
and instead am using a boolean returned from the Validate() call for
this purpose. The primary reason is that it seemed ridiculous to me to
implement an IsCompleted function on a struct aliased to a primitive
(like `OperatingState`) that just returned false
- Because of the additional `isValidated` private member, some tests
which previously deep compared object instances now fail. This is
because a test like `TestGetAddressable` will instantiate an Addressable,
pass it to core-metadata and then fetch the result. Upon umarshaling,
the result is validated whereas the original Addressable never has been.
Thus the `isValidated` members do not have the same value. It makes
more sense, therefore, to compare their exported values and so I am
now comparing the output of their respective String() operators.
- I am adopting the principle that aliased primitive structs (like
`AdminState` and `OperatingState`) must have their Validate() functions
invoked elsewhere and not by UnmarshalJSON.  This currently works
because these types are always supplied to the API as children of other
types. This will make more sense once the work on edgex-go issue 1244
is completed.
- As a basic guideline for what to validate, I am using the server-side
 logic found in the edgex-go controllers. This doesn't mean I'm
validating everything that could be validated, but I am trying to
achieve parity with what is there currently. We can add more as we go.
Also once this is integrated with edgex-go, a subsequent effort should
be undertaken to remove the server-side validation that will be
superfluous.
- I have removed the Service model since its only purpose was as a
composite used by DeviceService. There wasn't really any value in it.
- Some type instances used for testing have been unexported. There's no
reason for anything outside of their respective package to acces them.
There should be a subsequent clean-up issue to reconcile all of these.

Signed-off-by: Trevor Conn <trevor_conn@dell.com>