Fixed Event Struct Tags
This commit fixes issue 76

Updated the struct tags on the Event type. Previously, the id portion of
the struct tag was not specified for the "codec" type, which was causing
issues when encoding and decoding into CBOR using the codec library.

Added "codec" struct tags to the Reading type

Added tests for both types which validates CBOR encoding and decoding
Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>