Added AddBytes & MarshalEvent methods to EventClient
The Device Services group asked that the EventClient take on the
centralized responsibility for handling encoding of events. To that end,
marshaling has been exposed to callers via the MarshalEvent method.
After the caller uses this method to encode the event, they can add
the event directly as a byte array via the new EventClient.AddBytes
method.

Signed-off-by: Trevor Conn <trevor_conn@dell.com>