Removing Required Device Name from Reading
- Requirement for the Device name to be populated on a reading is being
removed per request from the Device Services group. The setting of a
device name on each reading is redundant when the readings are always
children of an event, and the event has the device name assigned to it.
Since readings can be queried independently of an event, however, we
will need a change in edgex-go to set this property at the time of
persistence from the event if it is not already set.

Signed-off-by: Trevor Conn <trevor_conn@dell.com>