Update Reading Validation-Allow Empty Binary Data (#241)
Fix #240

Update the validation of Reading to allow for empty `BinaryValue`. This
is needed since architecturally we are not persisting binary readings.
This means that Readings with a `ValueType` of `Binary` the
`BinaryValue` field is expected to be empty when obtaining existing
Readings. However, the `BinaryValue` field is still useful when creating
or updating an existing Reading as the data is expected to be populated.
Therefore, we must still enforce some validation such a the `MediaType`
being present when a `BinaryValue` is provided.

Update validation tests to accommodate the new changes and expectations.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>