feat(v2 api): Add new error kind for indicating requested range not satisfiable
As V2 API supports the pagination through offset and limit query string
when GET multiple objects, there is a need to add new error kind when
the requested range is not satisfiable. For example, when there are only
5 objects in the DB but users requests to retrieve objects starting from
10 (offset=9), V2 API shall return an error kind that such requested
range is not available. This new error kind shall also map to HTTP status
code 416 as indicated in the RFC 7233.

Signed-off-by: Jude Hung <jude@iotechsys.com>