feat(v2): refactor CoreCommand DTO to DeviceCoreCommand
The original CoreCommand DTO has field "DeviceName" to indicate which
device the command belongs to; however, the "DeviceName" becomes
redundant when query commands by deviceName, as each CoreCommand DTO
returned includes the same value of DeviceName.  As a result, this
commit refactor the CoreCommand DTO by extracting DeviceName up to a
level, so a single new DeviceCoreCommand DTO will have DeviceName and
ProfileName on the top level, and multiple CoreCommands in the bottom
level.

Signed-off-by: Jude Hung <jude@iotechsys.com>