feat!: Add interface for injecting authentication data to outgoing requests
BREAKING CHANGE: This change adds an AuthenticationInjector interface
to all http clients. If non-nil, the interface is invoked to add
authentication data to the outgoing HTTP request.

While the change does affect the public interface of the HTTP clients,
actual requests are only mutated by the authentication injector,
which is provided by the consumer.

Signed-off-by: Bryon Nevis <bryon.nevis@intel.com>